+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/NoPillaring.java
@@ -4,9 +4,12 @@ import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.ai.targeting.TargetingConditions;
+import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -21,6 +24,10 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 @LoadFeature(module = Modules.Ids.MOVEMENT)
 public class NoPillaring extends Feature {
 
+	@Config
+	@Label(name = "Monsters Only", description = "If true, pillaring will be negated only if there are monsters nearby")
+	public static Boolean monstersOnly = true;
+
 	public NoPillaring(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 	}
@@ -34,6 +41,9 @@ public class NoPillaring extends Feature {
 				|| playerEntity.isInWater()
 				|| playerEntity.onClimbable())
 			return;
+        if (monstersOnly
+				&& event.getLevel().getNearestEntity(Monster.class, TargetingConditions.forNonCombat().ignoreLineOfSight(), event.getEntity(), event.getEntity().getX(), event.getEntity().getY(), event.getEntity().getZ(), event.getEntity().getBoundingBox().inflate(16d)) == null)
+            return;
 		//noinspection ConstantConditions
 		BlockPos placedPos = event.getPos().relative(event.getFace());
 		Vec3 placedBlock = new Vec3(placedPos.getX() + 0.5d, placedPos.getY() + 0.5d, placedPos.getZ() + 0.5d);